keepass/requirements.txt
absl-py==0.9.0
argon2-cffi==19.2.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
cffi==1.13.2
construct==2.10.56
decorator==4.4.1
future==0.18.2
importlab==0.5.1
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
lxml==4.5.0
networkx==2.4
ninja==1.9.0.post1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
pykeepass==3.2.0
python-dateutil==2.8.1
PyYAML==5.3
six==1.14.0
soupsieve==1.9.5
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
